@@ -1,9 +1,9 @@
 {
   "name": "functions",
   "scripts": {
-    "build": "./node_modules/.bin/tslint -p tslint.json && ./node_modules/.bin/tsc",
+    "build": "eslint ./src && tsc -p .",
     "serve": "npm run build && firebase serve --only functions",
-    "shell": "npm run build && firebase experimental:functions:shell",
+    "shell": "npm run build && firebase emulators:start --only functions",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
@@ -20,29 +20,31 @@
     },
     {
       "name": "Lewis Freiberg"
+    },
+    {
+      "name": "Tim Sigl"
     }
   ],
   "dependencies": {
     "@iota/area-codes": "^0.1.1",
-    "@iota/bundle-validator": "1.0.0-beta.30",
-    "@iota/client-load-balancer": "github:iotaledger/client-load-balancer",
-    "@iota/converter": "1.0.0-beta.30",
-    "@iota/core": "1.0.0-beta.30",
-    "@iota/transaction-converter": "1.0.0-beta.30",
-    "axios": "^0.21.1",
+    "@iota/client": "^2.2.1",
+    "axios": "^0.24.0",
     "bluebird": "^3.7.2",
     "cors": "^2.8.5",
-    "firebase-admin": "^9.5.0",
-    "firebase-functions": "^3.13.1",
-    "mailgun-js": "^0.22.0"
+    "firebase-admin": "^10.0.0",
+    "firebase-functions": "^3.16.0",
+    "mailgun-js": "^0.22.0",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@firebase/app-types": "^0.6.1",
-    "tslint": "^6.1.3",
-    "typescript": "^3.8.3"
+    "@firebase/app-types": "^0.7.0",
+    "@typescript-eslint/eslint-plugin": "^5.3.0",
+    "@typescript-eslint/parser": "^5.3.0",
+    "eslint": "^8.1.0",
+    "typescript": "^4.4.4"
   },
   "engines": {
-    "node": "12"
+    "node": "14"
   },
   "resolutions": {
     "acorn": "^7.1.1",
