@@ -32,39 +32,40 @@
   },
   "private": true,
   "dependencies": {
-    "@iota/converter": "1.0.0-beta.30",
-    "@iota/core": "1.0.0-beta.30",
-    "@iota/http-client": "1.0.0-beta.30",
-    "@iota/mam": "^0.7.3",
+    "@iota/iota.js": "^1.8.2",
+    "@iota/mam.js": "^1.6.2",
     "date-fns": "^1.30.1",
     "file-saver": "^2.0.5",
     "firebase": "^8.2.7",
     "in-view": "^0.6.1",
-    "jszip": "^3.6.0",
-    "lodash-es": "^4.17.20",
-    "query-string": "^6.14.0",
-    "react": "^17.0.1",
-    "react-burger-menu": "^3.0.3",
+    "iota.crypto.js": "^0.4.2",
+    "jszip": "^3.7.1",
+    "lodash-es": "^4.17.21",
+    "mapbox-gl": "^2.6.0",
+    "query-string": "^7.0.1",
+    "react": "^17.0.2",
+    "react-burger-menu": "^3.0.6",
     "react-clipboard.js": "^2.0.16",
-    "react-cookie": "^4.0.3",
-    "react-dom": "^17.0.1",
+    "react-cookie": "^4.1.1",
+    "react-dom": "^17.0.2",
     "react-ga": "^3.3.0",
-    "react-map-gl": "^6.1.7",
+    "react-map-gl": "^6.1.17",
     "react-recaptcha": "^2.3.10",
-    "react-redux": "^7.2.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5",
+    "react-redux": "^7.2.6",
+    "react-router": "^5.2.1",
+    "react-router-dom": "^5.3.0",
+    "redux": "^4.1.2",
     "redux-logger": "^3.0.6",
     "redux-pack": "^0.1.5",
-    "styled-components": "5.2.1",
+    "styled-components": "5.3.3",
     "webfontloader": "^1.6.28"
   },
   "devDependencies": {
-    "@firebase/app-types": "^0.6.1",
+    "@firebase/app-types": "^0.7.0",
     "node-sass": "^5.0.0",
-    "react-scripts": "^4.0.2",
-    "webpack-bundle-analyzer": "^4.4.0"
+    "react-scripts": "^4.0.3",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "worker-loader": "^3.0.8"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -80,7 +81,7 @@
     "extends": "react-app"
   },
   "engines": {
-    "node": ">=10.0"
+    "node": "14"
   },
   "browserslist": [
     "> 2%",
@@ -89,14 +90,17 @@
   ],
   "resolutions": {
     "acorn": "^7.1.1",
+    "grpc": "^1.24.4",
     "immer": "^8.0.1",
     "ini": "^1.3.6",
     "kind-of": "^6.0.3",
-    "lodash": "^4.17.20",
+    "lodash": "^4.17.21",
     "minimist": "^1.2.5",
     "node-fetch": "^2.6.1",
     "node-forge": "^0.10.0",
     "node-notifier": "^8.0.1",
+    "postcss": "^7.0.36",
+    "tmpl": "^1.0.5",
     "websocket-extensions": "^0.1.4",
     "yargs-parser": "^18.1.3"
   }
