@@ -1,48 +1,36 @@
 import * as functions from 'firebase-functions';
 import * as admin from 'firebase-admin';
+import { Wallet } from './models/wallet';
+import { DeviceListItem } from './models/deviceListItem';
 
 admin.initializeApp(functions.config().firebase);
 
-exports.getKey = async (key: string) => {
+const getKey = async (key: string) => {
   // Get API key
-  const doc = await admin
-    .firestore()
-    .collection('keys')
-    .doc(key)
-    .get();
+  const doc = await admin.firestore().collection('keys').doc(key).get();
   if (doc.exists) return doc.data();
   console.error('getKey failed. API key is incorrect. ', key, doc);
   throw Error('Your API key is incorrect.');
 };
 
-exports.getSk = async (deviceId: string) => {
+const getSk = async (deviceId: string): Promise<DeviceListItem> => {
   // Get API key
-  const doc = await admin
-    .firestore()
-    .collection('deviceList')
-    .doc(deviceId)
-    .get();
-  if (doc.exists) return doc.data();
+  const doc = await admin.firestore().collection('deviceList').doc(deviceId).get();
+  if (doc.exists) return <DeviceListItem>doc.data();
   console.error('getSk failed. device does not exist', deviceId, doc);
   throw Error(`The device doesn't exist.`);
 };
 
-exports.getPurchase = async (uid: string, device: string) => {
+const getPurchase = async (uid: string, device: string) => {
   // Get User's purchase
-  const doc = await admin
-    .firestore()
-    .collection('users')
-    .doc(uid)
-    .collection('purchases')
-    .doc(device)
-    .get();
+  const doc = await admin.firestore().collection('users').doc(uid).collection('purchases').doc(device).get();
   // Check user's profile for purchase
   if (doc.exists) return doc.data();
   console.log('Device not purchased', uid, device);
   return false;
 };
 
-exports.getData = async (device: string, timestamp?: number) => {
+const getData = async (device: string, timestamp?: number) => {
   const time = timestamp ? Number(timestamp) : Date.now();
   // Get data
   const querySnapshot = await admin
@@ -58,7 +46,7 @@ exports.getData = async (device: string, timestamp?: number) => {
   if (querySnapshot.size === 0) return [];
 
   // Return data
-  return querySnapshot.docs.map(doc => {
+  return querySnapshot.docs.map((doc) => {
     if (doc.exists) {
       return doc.data();
     } else {
@@ -68,13 +56,9 @@ exports.getData = async (device: string, timestamp?: number) => {
   });
 };
 
-exports.getDevice = async (device: string, internal: boolean = false) => {
+const getDevice = async (device: string, internal = false) => {
   // Get User's purchase
-  const doc = await admin
-    .firestore()
-    .collection('devices')
-    .doc(device)
-    .get();
+  const doc = await admin.firestore().collection('devices').doc(device).get();
   // Check user's profile for purchase
   if (doc.exists) {
     const result = doc.data();
@@ -88,17 +72,14 @@ exports.getDevice = async (device: string, internal: boolean = false) => {
   return null;
 };
 
-exports.getDevices = async () => {
+const getDevices = async () => {
   // Get data
-  const querySnapshot = await admin
-    .firestore()
-    .collection('devices')
-    .get();
+  const querySnapshot = await admin.firestore().collection('devices').get();
 
   const promises = [];
   const results = [];
 
-  querySnapshot.docs.forEach(doc => {
+  querySnapshot.docs.forEach((doc) => {
     const promise = new Promise((resolve, reject) => {
       try {
         if (doc.exists) {
@@ -115,14 +96,14 @@ exports.getDevices = async () => {
             .collection('data')
             .limit(2)
             .get()
-            .then(deviceData => {
+            .then((deviceData) => {
               if (deviceData.size !== 0) {
                 result.hasData = true;
                 results.push(result);
-              };
+              }
               resolve(result);
             })
-            .catch(error => {
+            .catch((error) => {
               reject(error);
             });
         } else {
@@ -139,22 +120,18 @@ exports.getDevices = async () => {
   // Return data
   return await Promise.all(promises)
     .then(() => results)
-    .catch(error => {
+    .catch((error) => {
       console.log('getDevices error', error);
     });
 };
 
-exports.getUserDevices = async (user: string) => {
+const getUserDevices = async (user: string) => {
   // Get data
-  const querySnapshot = await admin
-    .firestore()
-    .collection('devices')
-    .where('owner', '==', user)
-    .get();
+  const querySnapshot = await admin.firestore().collection('devices').where('owner', '==', user).get();
   // Check there is data
   if (querySnapshot.size === 0) return [];
   // Return data
-  return querySnapshot.docs.map(doc => {
+  return querySnapshot.docs.map((doc) => {
     if (doc.exists) {
       const result = doc.data();
       delete result.sk;
@@ -167,37 +144,23 @@ exports.getUserDevices = async (user: string) => {
   });
 };
 
-exports.setPacket = async (device: string, packet: any) => {
+const setPacket = async (device: string, packet: any) => {
   // Save users API key and Seed
-  await admin
-    .firestore()
-    .collection('devices')
-    .doc(device)
-    .collection('data')
-    .doc()
-    .set(packet);
+  await admin.firestore().collection('devices').doc(device).collection('data').doc().set(packet);
 
   return true;
 };
 
-exports.setUser = async (uid: string, obj: any) => {
+const setUser = async (uid: string, obj: any) => {
   // Save users API key and Seed
-  await admin
-    .firestore()
-    .collection('users')
-    .doc(uid)
-    .set(obj);
+  await admin.firestore().collection('users').doc(uid).set(obj);
 
   return true;
 };
 
-exports.setDevice = async (deviceId: string, sk: string, address: string, seed: string, device: any) => {
+const setDevice = async (deviceId: string, sk: string, address: string, seed: string, device: any) => {
   // Save users API key and Seed
-  await admin
-    .firestore()
-    .collection('deviceList')
-    .doc(deviceId)
-    .set({ sk, seed });
+  await admin.firestore().collection('deviceList').doc(deviceId).set({ sk, seed });
 
   // Add public device record
   await admin
@@ -207,94 +170,49 @@ exports.setDevice = async (deviceId: string, sk: string, address: string, seed: 
     .set({ ...device, address });
 
   // Add device to owners' purchases
-  await admin
-    .firestore()
-    .collection('users')
-    .doc(device.owner)
-    .collection('purchases')
-    .doc(deviceId)
-    .set({
-      full: true,
-      time: Date.now(),
-    });
-
-  return true;
-};
-
-exports.setApiKey = async (apiKey: string, uid: string, email: string) => {
-  // Set API key in separate table
-  await admin
-    .firestore()
-    .collection('keys')
-    .doc(apiKey)
-    .set({
-      email,
-      uid,
-    });
-  return true;
-};
-
-exports.setOwner = async (deviceId: string, owner: string) => {
-  // Save new owner
-  await admin
-    .firestore()
-    .collection('devices')
-    .doc(deviceId)
-    .set({ owner }, { merge: true });
-  return true;
-};
-
-exports.setPurchase = async (userId: string, deviceId: string) => {
-  // Save new owner
-  await admin
-    .firestore()
-    .collection('users')
-    .doc(userId)
-    .collection('purchases')
-    .doc(deviceId)
-    .set({
-      full: true,
-      time: Date.now(),
-    });
-  return true;
-};
-
-exports.deleteDevice = async (device: any) => {
-  // Remove Device
-  await admin
-    .firestore()
-    .collection('deviceList')
-    .doc(device)
-    .delete();
-
-  // Get device data
-  const querySnapshot = await admin
-    .firestore()
-    .collection('devices')
-    .doc(device)
-    .collection('data')
-    .get();
-
-  querySnapshot.docs.forEach(async doc => {
-    // Delete device data
-    await admin
-      .firestore()
-      .collection('devices')
-      .doc(device)
-      .collection('data')
-      .doc(doc.id)
-      .delete();
+  await admin.firestore().collection('users').doc(device.owner).collection('purchases').doc(deviceId).set({
+    full: true,
+    time: Date.now()
   });
 
-  await admin
-    .firestore()
-    .collection('devices')
-    .doc(device)
-    .delete();
   return true;
 };
 
-exports.toggleWhitelistDevice = async (sensorId: string, inactive: string) => {
+const setApiKey = async (apiKey: string, uid: string, email: string) => {
+  // Set API key in separate table
+  await admin.firestore().collection('keys').doc(apiKey).set({
+    email,
+    uid
+  });
+  return true;
+};
+
+const setPurchase = async (userId: string, deviceId: string) => {
+  // Save new owner
+  await admin.firestore().collection('users').doc(userId).collection('purchases').doc(deviceId).set({
+    full: true,
+    time: Date.now()
+  });
+  return true;
+};
+
+const deleteDevice = async (device: any) => {
+  // Remove Device
+  await admin.firestore().collection('deviceList').doc(device).delete();
+
+  // Get device data
+  const querySnapshot = await admin.firestore().collection('devices').doc(device).collection('data').get();
+
+  querySnapshot.docs.forEach(async (doc) => {
+    // Delete device data
+    await admin.firestore().collection('devices').doc(device).collection('data').doc(doc.id).delete();
+  });
+
+  await admin.firestore().collection('devices').doc(device).delete();
+  return true;
+};
+
+const toggleWhitelistDevice = async (sensorId: string, inactive: string) => {
   // Whitelist device
   await admin
     .firestore()
@@ -304,14 +222,21 @@ exports.toggleWhitelistDevice = async (sensorId: string, inactive: string) => {
   return true;
 };
 
-exports.getUser = async (userId: string) => {
-  // Get user
-  const doc = await admin
-    .firestore()
-    .collection('users')
-    .doc(userId)
-    .get();
+const getDustProtectionThreshold = async (): Promise<number> => {
+  const doc = await admin.firestore().collection('settings').doc('settings').get();
+  if (doc.exists) {
+    const data = doc.data();
+    if (data.dustProtectionThreshold) {
+      return data.dustProtectionThreshold;
+    }
+  }
+  console.log('getDustProtectionThreshold failed. Setting does not exist', doc);
+  throw Error(`The getDustProtectionThreshold setting doesn't exist.`);
+};
 
+const getUser = async (userId: string) => {
+  // Get user
+  const doc = await admin.firestore().collection('users').doc(userId).get();
   // Check and return user
   if (doc.exists) {
     const result = doc.data();
@@ -329,12 +254,8 @@ exports.getUser = async (userId: string) => {
   return null;
 };
 
-exports.getNumberOfDevices = async () => {
-  const doc = await admin
-    .firestore()
-    .collection('settings')
-    .doc('settings')
-    .get();
+const getNumberOfDevices = async () => {
+  const doc = await admin.firestore().collection('settings').doc('settings').get();
   if (doc.exists) {
     const data = doc.data();
     if (data.numberOfDevices) {
@@ -345,13 +266,9 @@ exports.getNumberOfDevices = async () => {
   throw Error(`The getNumberOfDevices setting doesn't exist.`);
 };
 
-exports.getSettings = async () => {
+const getSettings = async () => {
   // Get data
-  const doc = await admin
-    .firestore()
-    .collection('settings')
-    .doc('settings')
-    .get();
+  const doc = await admin.firestore().collection('settings').doc('settings').get();
   if (doc.exists) {
     const {
       defaultPrice,
@@ -363,7 +280,8 @@ exports.getSettings = async () => {
       recaptchaSiteKey,
       tangleExplorer,
       nodes,
-      tangle
+      tangle,
+      dustProtectionThreshold
     } = doc.data();
     return {
       defaultPrice,
@@ -375,30 +293,26 @@ exports.getSettings = async () => {
       recaptchaSiteKey,
       tangleExplorer,
       nodes,
-      tangle
+      tangle,
+      dustProtectionThreshold
     };
   }
   console.error('getSettings failed. Setting does not exist', doc);
   throw Error(`The getSettings setting doesn't exist.`);
 };
 
-exports.getUserWallet = async (uid: string) => {
+const getUserWallet = async (uid: string): Promise<Wallet> => {
   // Get User's wallet
-  const doc = await admin
-    .firestore()
-    .collection('users')
-    .doc(uid)
-    .get();
+  const doc = await admin.firestore().collection('users').doc(uid).get();
 
   if (doc.exists) {
     const data = doc.data();
     return data.wallet || null;
   }
-  console.log('getUserWallet failed. ', uid);
-  throw Error(`The wallet doesn't exist.`);
+  return null;
 };
 
-exports.setWallet = async (uid: string, wallet: any) => {
+const setWallet = async (uid: string, wallet: any) => {
   // Create wallet
   await admin
     .firestore()
@@ -408,30 +322,13 @@ exports.setWallet = async (uid: string, wallet: any) => {
   return true;
 };
 
-exports.updateBalance = async (uid: string, balance: any) => {
-  await admin
-    .firestore()
-    .collection('users')
-    .doc(uid)
-    .set({ wallet: { balance } }, { merge: true });
+const updateBalance = async (uid: string, balance: any) => {
+  await admin.firestore().collection('users').doc(uid).set({ wallet: { balance } }, { merge: true });
   return true;
 };
 
-exports.updateUserWalletAddressKeyIndex = async (address: string, keyIndex: number, uid: string) => {
-  await admin
-    .firestore()
-    .collection('users')
-    .doc(uid)
-    .set({ wallet: { address, keyIndex } }, { merge: true });
-  return true;
-};
-
-exports.getIotaWallet = async () => {
-  const doc = await admin
-    .firestore()
-    .collection('settings')
-    .doc('settings')
-    .get();
+const getIotaWallet = async () => {
+  const doc = await admin.firestore().collection('settings').doc('settings').get();
   if (doc.exists) {
     const data = doc.data();
     if (data.wallet) {
@@ -442,21 +339,8 @@ exports.getIotaWallet = async () => {
   throw Error(`The getIotaWallet setting doesn't exist.`);
 };
 
-exports.updateWalletAddressKeyIndex = async (address: string, keyIndex: number, userId: string) => {
-  await admin
-    .firestore()
-    .collection('settings')
-    .doc('settings')
-    .set({ wallet: { address, keyIndex } }, { merge: true });
-  return true;
-};
-
-exports.getEmailSettings = async () => {
-  const doc = await admin
-    .firestore()
-    .collection('settings')
-    .doc('settings')
-    .get();
+const getEmailSettings = async () => {
+  const doc = await admin.firestore().collection('settings').doc('settings').get();
   if (doc.exists) {
     const data = doc.data();
     return data.email || null;
@@ -465,12 +349,8 @@ exports.getEmailSettings = async () => {
   throw Error(`The getEmailSettings setting doesn't exist.`);
 };
 
-exports.getGoogleMapsApiKey = async () => {
-  const doc = await admin
-    .firestore()
-    .collection('settings')
-    .doc('settings')
-    .get();
+const getGoogleMapsApiKey = async () => {
+  const doc = await admin.firestore().collection('settings').doc('settings').get();
   if (doc.exists) {
     const data = doc.data();
     if (data.googleMapsApiKey) {
@@ -479,4 +359,31 @@ exports.getGoogleMapsApiKey = async () => {
   }
   console.log('getGoogleMapsApiKey failed. Setting does not exist', doc);
   throw Error(`The getGoogleMapsApiKey setting doesn't exist.`);
+};
+
+export {
+  getKey,
+  getSk,
+  getPurchase,
+  getData,
+  getDevice,
+  getDevices,
+  getUserDevices,
+  setPacket,
+  setUser,
+  setDevice,
+  setApiKey,
+  setPurchase,
+  deleteDevice,
+  toggleWhitelistDevice,
+  getUser,
+  getNumberOfDevices,
+  getSettings,
+  getUserWallet,
+  setWallet,
+  updateBalance,
+  getIotaWallet,
+  getEmailSettings,
+  getGoogleMapsApiKey,
+  getDustProtectionThreshold
 };
