@@ -1,8 +1,7 @@
 import * as functions from 'firebase-functions';
 const cors = require('cors')({ origin: true });
-const { validateBundleSignatures } = require('@iota/bundle-validator');
 
-const {
+import {
   getKey,
   getSk,
   getPurchase,
@@ -18,22 +17,19 @@ const {
   setDevice,
   setPacket,
   setPurchase,
-  setOwner,
   setApiKey,
   setWallet,
   deleteDevice,
   toggleWhitelistDevice,
   updateBalance,
-  updateUserWalletAddressKeyIndex,
   getEmailSettings,
-} = require('./firebase');
+  getDustProtectionThreshold
+} from './firebase';
 const { sendEmail } = require('./email');
-const {
+import {
   generateUUID,
   generateSeed,
-  generateNewAddress,
   sanatiseObject,
-  findTx,
   faucet,
   initWallet,
   checkRecaptcha,
@@ -42,7 +38,32 @@ const {
   addressToIac,
   gpsToIac,
   initSemarketWallet,
-} = require('./helpers');
+  getBalance,
+  fundWallet
+} from './helpers';
+
+exports.balance = functions.https.onRequest((req, res) => {
+  cors(req, res, async () => {
+    const params = req.query;
+    if (!params && !params.address) {
+      console.error('balance failed: ', params);
+      return res.status(400).json({ error: 'Ensure all fields are included' });
+    }
+    try {
+      const address = <string>params.address;
+      const balanceWithDustProtection = await getBalance(address);
+      const dustProtectionThreshold = await getDustProtectionThreshold();
+      const balance =
+        balanceWithDustProtection - dustProtectionThreshold >= 0
+          ? balanceWithDustProtection - dustProtectionThreshold
+          : 0;
+      return res.status(200).json(balance);
+    } catch (error) {
+      console.error('Could not get balance', error);
+      return res.status(500).json({ error: 'Could not get balance' });
+    }
+  });
+});
 
 // Take in data from device
 exports.newData = functions.https.onRequest((req, res) => {
@@ -58,7 +79,7 @@ exports.newData = functions.https.onRequest((req, res) => {
       const device = await getSk(packet.id);
       if (device.sk === packet.sk) {
         return res.json({
-          success: await setPacket(packet.id, packet.packet),
+          success: await setPacket(packet.id, packet.packet)
         });
       } else {
         console.error('newData failed. Key is incorrect', device.sk, packet.sk);
@@ -86,22 +107,25 @@ exports.newDevice = functions.https.onRequest((req, res) => {
 
     try {
       const invalid = sanatiseObject(packet.device);
-      const secretKey = generateSeed(15);
-      const seed = generateSeed();
-      const address = generateNewAddress(seed);
+      const secretKey = generateSeed();
       if (invalid) throw Error(invalid);
 
-      const key = await getKey(<String>packet.apiKey);
+      const key = await getKey(<string>packet.apiKey);
       const userDevices = await getUserDevices(key.uid);
       const user = await getUser(key.uid);
       if (!user.numberOfDevices) {
         user.numberOfDevices = await getNumberOfDevices();
       }
       if (userDevices.length < user.numberOfDevices) {
-        const device = await getDevice(<String>packet.id, true);
+        const device = await getDevice(<string>packet.id, true);
         if (device && device.owner !== key.uid) {
-          return res.json({ error: `Device with ID ${packet.id} already exists. Please specify new unique ID` });
+          return res.json({
+            error: `Device with ID ${packet.id} already exists. Please specify new unique ID`
+          });
         }
+        const {
+          wallet: { address, seed }
+        } = await initWallet();
 
         return res.json({
           success: await setDevice(packet.id, secretKey, address, seed, packet.device),
@@ -109,7 +133,9 @@ exports.newDevice = functions.https.onRequest((req, res) => {
         });
       } else {
         console.error('newDevice failed. You have too many devices', userDevices.length);
-        return res.json({ error: 'You have too many devices. Please delete one to clear space' });
+        return res.json({
+          error: 'You have too many devices. Please delete one to clear space'
+        });
       }
     } catch (e) {
       console.error('newDevice failed. Error: ', e.message);
@@ -130,28 +156,24 @@ exports.delete = functions.https.onRequest((req, res) => {
 
     try {
       const { apiKey, deviceId } = packet;
-      const key = await getKey(<String>apiKey);
-      const device = await getDevice(<String>deviceId, true);
+      const key = await getKey(<string>apiKey);
+      const device = await getDevice(<string>deviceId, true);
 
       if (!device) {
         throw Error(`Device doesn't exist`);
       }
       if (device.owner === key.uid) {
         return res.json({
-          success: await deleteDevice(<String>deviceId),
+          success: await deleteDevice(<string>deviceId)
         });
       } else {
-        console.error(
-          "removeDevice failed. You don't have permission to delete this device",
-          device.owner,
-          key.uid
-        );
+        console.error("removeDevice failed. You don't have permission to delete this device", device.owner, key.uid);
         throw Error(`You don't have permission to delete this device`);
       }
     } catch (e) {
       console.error('removeDevice failed. Error: ', e.message);
       return res.status(403).json({
-        error: e.message,
+        error: e.message
       });
     }
   });
@@ -163,10 +185,11 @@ exports.devices = functions.https.onRequest((req, res) => {
     try {
       const params = req.query;
       if (params && params.userId && params.apiKey) {
-        const { uid } = await getKey(<String>params.apiKey);
+        const { uid } = await getKey(<string>params.apiKey);
         if (params.userId === uid) {
-          const userDevices = await getUserDevices(params.userId);
-          const promises = await userDevices.map(async device => {
+          const userDevices = await getUserDevices(params.userId as string);
+          const promises = await userDevices.map(async (device) => {
+            // eslint-disable-next-line no-async-promise-executor
             const promise = await new Promise(async (resolve, reject) => {
               try {
                 const keyObj = await getSk(device.sensorId);
@@ -206,7 +229,7 @@ exports.device = functions.https.onRequest((req, res) => {
     }
 
     try {
-      const device = await getDevice(params.deviceId);
+      const device = await getDevice(params.deviceId as string);
       if (!device) {
         throw Error(`Device doesn't exist`);
       }
@@ -233,12 +256,12 @@ exports.stream = functions.https.onRequest((req, res) => {
 
     try {
       // Make sure purchase exists
-      const purchase = await getPurchase(<String>params.userId, <String>params.deviceId);
+      const purchase = await getPurchase(<string>params.userId, <string>params.deviceId);
       if (!purchase) {
         return res.json({ success: false });
       }
       // Return data
-      return res.json(await getData(<String>params.deviceId, params.time));
+      return res.json(await getData(<string>params.deviceId, params.time as any));
     } catch (e) {
       console.error('stream failed. Error: ', e.message);
       return res.status(403).json({ error: e.message });
@@ -247,8 +270,9 @@ exports.stream = functions.https.onRequest((req, res) => {
 });
 
 // // Setup User with an API Key
-exports.setupUser = functions.auth.user().onCreate(user => {
-  return new Promise(async (resolve, reject) => {
+exports.setupUser = functions.auth.user().onCreate((user) => {
+  // eslint-disable-next-line no-async-promise-executor
+  return new Promise(async (resolve, reject): Promise<void> => {
     if (!user.email) {
       reject();
     } else {
@@ -256,12 +280,9 @@ exports.setupUser = functions.auth.user().onCreate(user => {
       try {
         const apiKey = generateUUID();
         const numberOfDevices = (await getNumberOfDevices()) || 5;
-
         await setUser(user.uid, { apiKey, numberOfDevices });
         await setApiKey(apiKey, user.uid, user.email);
-
-        console.log('setupUser resolved for UID', user.uid);
-        resolve();
+        resolve(null);
       } catch (e) {
         console.error('setupUser rejected with ', e.message);
         reject(e.message);
@@ -281,7 +302,7 @@ exports.toggleWhitelist = functions.https.onRequest((req, res) => {
     }
 
     try {
-      const data = await getKey(<String>packet.apiKey);
+      const data = await getKey(<string>packet.apiKey);
       if (data.email && data.email.indexOf('iota.org') !== -1 && packet.uid === data.uid) {
         // Toggle whitelist
         console.error('toggleWhitelist success', packet, data);
@@ -307,7 +328,7 @@ exports.user = functions.https.onRequest((req, res) => {
 
     try {
       // Retrieve user
-      const user = await getUser(params.userId);
+      const user = await getUser(params.userId as string);
       if (!user) {
         return res.json(null);
       }
@@ -371,11 +392,19 @@ exports.wallet = functions.https.onRequest((req, res) => {
       console.error('wallet failed. Packet: ', packet);
       return res.status(400).json({ error: 'Malformed Request' });
     }
-
+    const userId = packet.userId;
     try {
-      const result = await initWallet(packet.userId);
-      await setWallet(packet.userId, result.wallet);
-      return res.json({ success: result.transactions.length > 0 });
+      const wallet = await getUserWallet(userId);
+      // If wallet already exits, refill wallet
+      if (wallet) {
+        const filledWallet = await fundWallet(wallet.address, wallet.seed);
+        await setWallet(packet.userId, filledWallet.wallet);
+        return res.status(200).json({ messageId: filledWallet.messageId });
+      } else {
+        const initializedWallet = await initWallet();
+        await setWallet(packet.userId, initializedWallet.wallet);
+        return res.status(201).json({ messageId: initializedWallet.messageId });
+      }
     } catch (e) {
       console.error('wallet failed. Error: ', e.message);
       return res.status(403).json({ error: e.message });
@@ -401,8 +430,8 @@ exports.faucet = functions.https.onRequest((req, res) => {
         return res.status(403).json({ error: recaptcha['error-codes'] });
       }
 
-      const transactions = await faucet(packet.address);
-      return res.json({ transactions });
+      const { messageId } = await faucet(packet.address);
+      return res.json({ messageId });
     } catch (e) {
       console.error('faucet failed. Error: ', e.message);
       return res.status(403).json({ error: e.message });
@@ -422,7 +451,7 @@ exports.purchaseStream = functions.https.onRequest((req, res) => {
     try {
       const device = await getDevice(packet.deviceId);
       const wallet = await getUserWallet(packet.userId);
-      const { iotaApiVersion, nodes, defaultPrice } = await getSettings();
+      const { defaultPrice } = await getSettings();
       let price = defaultPrice;
       if (device) {
         if (device.price) {
@@ -433,39 +462,23 @@ exports.purchaseStream = functions.https.onRequest((req, res) => {
       } else {
         return res.json({ error: `Device doesn't exist` });
       }
-
-      let newWalletBalance;
-      if (wallet && wallet.balance) {
-        newWalletBalance = Number(wallet.balance) - Number(price);
-        if (newWalletBalance < 0) {
-          console.error('purchaseStream failed. Not enough funds', packet);
-          return res.json({ error: 'Not enough funds or your new wallet is awaiting confirmation. Please try again in 5 min.' });
-        }
-      } else {
-        console.error('purchaseStream failed. Wallet not set', packet);
-        return res.json({ error: 'Wallet not set' });
+      const dustProtectionThreshold = await getDustProtectionThreshold();
+      const newWalletBalance = Number(wallet.balance) - Number(price);
+      if (newWalletBalance < dustProtectionThreshold) {
+        console.error('purchaseStream failed. Not enough funds', packet);
+        return res.json({
+          error: 'Not enough funds in your wallet.'
+        });
+      }
+      try {
+        await purchaseData(packet.userId, packet.deviceId, device.address, price);
+      } catch (e) {
+        return res.status(403).json({ error: e.message });
       }
 
-      const transactions = await purchaseData(packet.userId, device.address, price);
-      console.log('purchaseStream', packet.userId, packet.deviceId, transactions);
-
-      if (transactions) {
-        const hashes = transactions && transactions.map(transaction => transaction.hash);
-
-        // Find TX on network and parse
-        const bundle = await findTx(hashes, nodes[0], iotaApiVersion);
-
-        // Make sure TX is valid
-        if (!validateBundleSignatures(bundle)) {
-          console.error('purchaseStream failed. Transaction is invalid for: ', bundle);
-          res.status(403).json({ error: 'Transaction is Invalid' });
-        }
-
-        await setPurchase(packet.userId, packet.deviceId);
-        await updateBalance(packet.userId, newWalletBalance);
-        return res.json({ success: true });
-      }
-      return res.json({ error: 'Purchase failed. Insufficient balance of out of sync' });
+      await setPurchase(packet.userId, packet.deviceId);
+      await updateBalance(packet.userId, newWalletBalance);
+      return res.json({ success: true });
     } catch (e) {
       console.error('purchaseData failed. Error: ', e, packet);
       return res.status(403).json({ error: e.message });
@@ -478,8 +491,8 @@ exports.semarket = functions.https.onRequest((req, res) => {
     try {
       const params = req.query;
       if (params.address) {
-        const transactions = await initSemarketWallet(params.address, params.amount || null);
-        return res.json({ success: transactions.length > 0 });
+        const messageId = await initSemarketWallet(<string>params.address, parseInt(<string>params.amount) || null);
+        return res.json({ messageId: messageId });
       }
       return res.json({ success: false, error: 'no address' });
     } catch (e) {
@@ -496,14 +509,14 @@ exports.location = functions.https.onRequest((req, res) => {
       let result = null;
       if (params.address) {
         result = await addressToIac(decodeURI(params.address.toString()));
-        console.log(`Converted address "${decodeURI(params.address.toString())}" to "${result}"`);
       } else if (params.iac) {
         result = await iacToAddress(params.iac);
-        console.log(`Converted area code "${params.iac}" to "${result}"`);
       } else if (params.gps) {
-        const coordinates = params.gps.toString().split(',').map(coord => Number(coord));
-        result = await gpsToIac(...coordinates);
-        console.log(`Converted GPS coordinates "${params.gps}" to "${result}"`);
+        const coordinates = params.gps
+          .toString()
+          .split(',')
+          .map((coord) => Number(coord));
+        result = await gpsToIac(...(coordinates as [number, number]));
       }
       return res.json(result);
     } catch (e) {
