@@ -89,8 +89,7 @@ class Form extends React.Component {
     this.setState(
       {
         [name]: value,
-      },
-      () => console.log(this.state)
+      }
     );
   }
 
