@@ -2,12 +2,15 @@ import React, { useContext, useEffect } from 'react';
 import InView from 'in-view';
 import { SensorContext } from '../../pages/sensor';
 
-export default ({ children }) => {
+const InViewEvent = ({ children }) => {
   const { func } = useContext(SensorContext);
 
+  // Attach eventListener to last data packet to load new data when last packet is in view
   useEffect(() => {
-    InView('.inview').on('enter', el => func());
+    InView('.inview').on('enter', (el) => func());
   }, []);
 
   return <div className={'inview'}>{children}</div>;
-}
+};
+
+export default InViewEvent;
