@@ -1,12 +1,15 @@
 import React from 'react';
 import ReactGA from 'react-ga';
-import styled from 'styled-components';
-import MapGL, { Popup, NavigationControl } from 'react-map-gl';
-import { Link } from 'react-router-dom';
+import MapGL, { Popup, NavigationControl } from '!react-map-gl'; // eslint-disable-line import/no-webpack-loader-syntax
 import { connect } from 'react-redux';
-import '../../assets/scss/mapbox.scss';
 import Controls from './controls';
 import Markers from './markers';
+import 'mapbox-gl/dist/mapbox-gl.css';
+import { CardFooter, CardHeader, FootRow, Header, HeaderBg, HeaderBgMobile, Heading, InfoKey, InfoValue, LocationIcon, Main, SensorCard, SensorId, SensorType, Subtitle } from './styles';
+import 'mapbox-gl/dist/mapbox-gl.css';
+import mapboxgl from 'mapbox-gl';
+// eslint-disable-next-line import/no-webpack-loader-syntax
+mapboxgl.workerClass = require('worker-loader!mapbox-gl/dist/mapbox-gl-csp-worker').default;
 
 const mapControls = new Controls();
 
@@ -134,41 +137,41 @@ class Map extends React.Component {
     if (!settings.mapboxApiAccessToken) return <div />;
 
     return (
-      <Main id="map">
-        <Header>
-          <div>
-            <Heading>Sensor map</Heading>
-            <Subtitle>Click on a pin to view the sensor information.</Subtitle>
-          </div>
-        </Header>
-        <MapGL
-          scrollZoom={false}
-          controller={mapControls}
-          maxZoom={11.5}
-          {...viewport}
-          height={mapHeight}
-          mapStyle={settings.mapboxStyles}
-          onViewportChange={this.updateViewport}
-          onClick={() => (popupInfo ? this.setState({ popupInfo: null }) : null)}
-          mapboxApiAccessToken={settings.mapboxApiAccessToken}>
-          <div style={{ position: 'absolute', right: 20, top: 10 }}>
-            <NavigationControl onViewportChange={this.updateViewport} />
-          </div>
-          <Markers devices={devices} openPopup={this.openPopup} />
-          {this.renderPopup()}
-        </MapGL>
+<Main id="map">
+  <Header>
+    <div>
+      <Heading>Sensor map</Heading>
+      <Subtitle>Click on a pin to view the sensor information.</Subtitle>
+    </div>
+  </Header>
+  <MapGL
+    scrollZoom={false}
+    controller={mapControls}
+    maxZoom={11.5}
+    {...viewport}
+    height={mapHeight}
+    mapStyle='mapbox://styles/mapbox/streets-v11'
+    onViewportChange={this.updateViewport}
+    onClick={() => (popupInfo ? this.setState({ popupInfo: null }) : null)}
+    mapboxApiAccessToken={settings.mapboxApiAccessToken}>
+<div style={{ position: 'absolute', right: 60, top: 10 }}>
+  <NavigationControl onViewportChange={this.updateViewport} />
+</div>
+<Markers devices={devices} openPopup={this.openPopup} />
+{this.renderPopup()}
+  </MapGL>
 
-        <HeaderBg
-          src="/static/shapes/shape-main-1.svg"
-          className="shape-accent-1 mobile-hidden"
-          alt="Shape accent"
-        />
-        <HeaderBgMobile
-          src="/static/shapes/shape-main-1-mobile.svg"
-          className="shape-accent-1 desktop-hidden shape-mobile"
-          alt="Shape accent"
-        />
-      </Main>
+  <HeaderBg
+    src="/static/shapes/shape-main-1.svg"
+    className="shape-accent-1 mobile-hidden"
+    alt="Shape accent"
+  />
+  <HeaderBgMobile
+    src="/static/shapes/shape-main-1-mobile.svg"
+    className="shape-accent-1 desktop-hidden shape-mobile"
+    alt="Shape accent"
+  />
+</Main>
     );
   }
 }
@@ -178,180 +181,3 @@ const mapStateToProps = state => ({
 });
 
 export default connect(mapStateToProps)(Map);
-
-const Main = styled.div`
-  position: relative;
-  &::before {
-    content: '';
-    position: absolute;
-    top: 74px;
-    left: 0;
-    height: 1px;
-    width: 100vw;
-    background-color: #eaecee;
-    @media (max-width: 760px) {
-      top: 0;
-    }
-  }
-`;
-
-const Header = styled.div`
-  max-width: 1170px;
-  padding: 0 15px;
-  margin-right: auto;
-  margin-left: auto;
-  display: flex;
-  justify-contents: center;
-  width: 300px;
-  position: absolute;
-  z-index: 900;
-  padding-top: 140px;
-  @media (max-width: 760px) {
-    padding-top: 50px;
-  }
-  @media (max-width: 520px) {
-    width: 200px;
-  }
-  @media (max-width: 400px) {
-    padding-top: 0;
-    margin-top: -100px;
-  }
-`;
-
-const Heading = styled.h3`
-  text-transform: capitalize;
-  text-align: left;
-  color: #009fff;
-  font-size: 28px;
-  font-weight: 100;
-  line-height: 42px;
-  margin-bottom: 12px;
-  @media (max-width: 760px) {
-    font-size: 24px;
-    margin-bottom: 0;
-    margin-bottom: 10px;
-  }
-`;
-
-const Subtitle = styled.h4`
-  font-size: 19px;
-  font-weight: 400;
-  line-height: 33px;
-  margin-bottom: 60px;
-  max-width: 265px;
-  color: #fff;
-  text-align: left;
-  @media (max-width: 1120px) {
-    max-width: 215px;
-  }
-  @media (max-width: 760px) {
-    font-size: 18px;
-    line-height: 28px;
-  }
-`;
-
-const HeaderBg = styled.img`
-  position: absolute;
-  top: -120px;
-  right: 75vw;
-  z-index: 800;
-  @media (max-width: 1120px) {
-    right: 67vw;
-  }
-  @media (max-width: 760px) {
-    display: none;
-  }
-`;
-
-const HeaderBgMobile = styled.img`
-  position: absolute;
-  top: -20px;
-  right: 70vw;
-  z-index: 800;
-  display: none;
-  @media (max-width: 1120px) {
-    right: 67vw;
-  }
-  @media (max-width: 760px) {
-    top: -100px;
-    right: 55vw;
-    display: block;
-  }
-  @media (max-width: 400px) {
-    top: -250px;
-    right: 50vw;
-  }
-`;
-
-const SensorCard = styled(Link)`
-  display: block;
-  border-radius: 6px;
-  transition: box-shadow 0.19s ease-out;
-  position: relative;
-  user-select: none;
-  cursor: default;
-  color: inherit;
-  text-decoration: none;
-  width: 280px;
-  height: 150px;
-  padding-top: 19px;
-  border: none;
-  background-color: #0e38a0;
-  box-shadow: 0 14px 28px 0 rgba(10, 32, 87, 0.24);
-  @media (max-width: 760px) {
-    margin-bottom: 10px;
-  }
-`;
-
-const CardHeader = styled.header`
-  position: relative;
-  padding: 0 30px 8px 30px;
-  border-bottom: 1px solid rgba(115, 143, 212, 0.2);
-`;
-
-const CardFooter = styled.footer`
-  padding: 20px 30px;
-  padding: 8px 30px;
-  border-top: none;
-  background-color: transparent;
-`;
-
-const FootRow = styled.div`
-  display: flex;
-  justify-content: space-between;
-  &:not(:last-of-type) {
-    margin-bottom: 5px;
-  }
-`;
-
-const InfoKey = styled.span`
-  color: #738fd4;
-  text-transform: capitalize;
-  font: 12px/16px 'Nunito Sans', sans-serif;
-`;
-
-const InfoValue = styled.span`
-  font-size: 12px;
-  line-height: 16px;
-  font-weight: 400;
-  color: #fff;
-`;
-
-const LocationIcon = styled.img`
-  margin: 0 6px 0 13px;
-`;
-
-const SensorType = styled.span`
-  font: 12px/16px 'Nunito Sans', sans-serif;
-  position: absolute;
-  top: -8px;
-  color: #738fd4;
-`;
-
-const SensorId = styled.span`
-  font-size: 20px;
-  top: 4px;
-  color: #fff;
-  line-height: 42px;
-  position: relative;
-`;
