@@ -11,7 +11,7 @@ export const initializeFirebaseApp = () => {
 
 export const userAuth = async () => {
   return new Promise((resolve, reject) => {
-    firebase.auth().onAuthStateChanged(async user => {
+    firebase.auth().onAuthStateChanged(async (user) => {
       if (user) {
         // User is signed in.
         resolve(user);
@@ -19,13 +19,13 @@ export const userAuth = async () => {
         firebase
           .auth()
           .signInAnonymously()
-          .then(data => {
+          .then((data) => {
             if (data.user) {
               resolve(data.user);
             }
             reject(data);
           })
-          .catch(error => {
+          .catch((error) => {
             console.error('userAuth error', error);
             reject(error.message);
             return error.message;
@@ -37,8 +37,12 @@ export const userAuth = async () => {
 
 export const allDevices = () => {
   return new Promise(async (resolve, reject) => {
-    const devices = await api.get('devices');
-    const devicesWithData = devices.filter(device => device.hasData);
-    resolve(devicesWithData);
+    try {
+      const devices = await api.get('devices');
+      const devicesWithData = devices.filter((device) => device.hasData);
+      resolve(devicesWithData);
+    } catch (error) {
+      reject(error);
+    }
   });
 };
